Implement encoder/ decoder methods for Tokenizer.
For encoding: parse string into inidivual byte codes for which we iteratively apply the original merges and then map to the learned vocabulary of tokens.

For decoding: simply iterate over the provided tokens and fetch the byte code from the learned vocabulary and parse to UTF-8 Unicode so that it is human readable.