Implement FeedForward using SwiGLU algorithm.
The SwiGLU (SiLU/Swish GLU) is a combination of the SiLU activation sandwiched in a Gated-Linear-Unit operation. GLU's are suggested to reduce the vanishing gradient problem for deep architectures by providing a linear path for the gradients while retaining non-linear capabilities.

Added an __init__.py file to the modules folder to simplify module imports.