Implement numerically stable softmax.
Softmax can become infinite for large values (then inf/inf = NaN). To avoid this we can notice that the softmax op is invariant to adding any constant to all inputs; we can leverage this property for numeric stability: subtract the largest entry of the list from all elements making the new largest entry 0.