Implement DataLoader service.
To load information from raw bytes, tokenize it and group it into batches for training. I had to move the tokenizer implementation into the `utils/` folder to simplify imports. I am also not checking for <|endoftext|> tokens in the middle of examples which will likely affect model training but I want to also fix the file reading API in a unified way to also simplify the tokenizer impl. but I'd rather send that fix in a follow up CL; for now this should suffice.