Move Tokenizer out from Config object to simplify pickling.
When saving the pickle binary using torch.save() we can't simply store arbitrary Python classes due to lack of parsing guidance when loading them back in memory (ie torch.load()). I am removing the Tokenizer instance from the Config class and managing that separately instead so that we don't have to "serialize" it or special-case it. This simplifies the checkpointing infra quite a bit.