Train model for 16_600 steps.
The performance seems pretty reasonable for the current model size. I trained for 16k+ steps over 2-3 days on a single MPS chip achieving a minimum cross entropy loss of ~2.3 which produces fairly legible sentences from the TinyStories dataset. This is a great POC that the current infrastructure is capable of actually producing functional models.